Add checks while removing project members

+++ packages/backend/src/resolvers.ts
@@ -1,7 +1,8 @@
 import debug from 'debug';
+import assert from 'assert';
 
 import { Database } from './database';
-import { deploymentToGqlType, projectMemberToGqlType, projectToGqlType, environmentVariableToGqlType } from './utils';
+import { deploymentToGqlType, projectMemberToGqlType, projectToGqlType, environmentVariableToGqlType, isUserOwner } from './utils';
 
 const log = debug('snowball:database');
 
@@ -69,20 +70,33 @@ export const createResolvers = async (db: Database): Promise<any> => {
     },
 
     Mutation: {
-      removeMember: async (_: any, { memberId }: { memberId: string }) => {
+      removeMember: async (_: any, { memberId }: { memberId: string }, context: any) => {
         try {
-          return db.removeProjectMemberByMemberId(memberId);
-        } catch (error) {
-          log(error);
+          const member = await db.getProjectMemberByMemberId(memberId);
+
+          if (member.member.id === context.userId) {
+            throw new Error('Invalid operation: cannot remove self');
+          }
+
+          const memberProject = member.project;
+          assert(memberProject);
+
+          if (isUserOwner(String(context.userId), String(memberProject.owner.id))) {
+            return db.removeProjectMemberByMemberId(memberId);
+          } else {
+            throw new Error('Invalid operation: not authorized');
+          }
+        } catch (err) {
+          log(err);
           return false;
         }
       },
 
-      addEnvironmentVariables: async (_: any, { projectId, environmentVariables }: { projectId: string, environmentVariables: any[] }) => {
+      addEnvironmentVariables: async (_: any, { projectId, environmentVariables }: { projectId: string, environmentVariables: { environments: string[], key: string, value: string}[] }) => {
         try {
           return db.addEnvironmentVariablesByProjectId(projectId, environmentVariables);
-        } catch (error) {
-          log(error);
+        } catch (err) {
+          log(err);
           return false;
         }
       }
